refactor: 폴더 이동 (mock 객체는 mock으로 따로 분리)
가독성 증진을 위해 mock 폴더 분리
